+++ backend/base/serializers.py
@@ -9,7 +9,7 @@ class CustomAnalyticsSerializer(serializers.Serializer):
     """
     age = serializers.IntegerField(required=False)
     ethnicity = serializers.CharField(required=False)
-    tumor_grade = serializers.IntegerField(required=False)
+    tumor_grade = serializers.FloatField(required=False)
     tumor_size = serializers.CharField(required=False)
     num_pos_nodes = serializers.CharField(required=False)
     er_status = serializers.CharField(required=False)
@@ -26,7 +26,7 @@ class DiagnosisSerializer(serializers.Serializer):
     """
     age = serializers.IntegerField(required=True)
     tumor_size_in_mm = serializers.IntegerField(required=False, default=0)
-    tumor_grade = serializers.IntegerField(required=True)
+    tumor_grade = serializers.FloatField(required=True)
     er_status = serializers.CharField(required=True)
     pr_status = serializers.CharField(required=True)
     her2_status = serializers.CharField(required=True)
